Combined metrics for same model as boxplot for completed distance

+++ behavior_metrics/driver_carla.py
@@ -219,6 +219,7 @@ def generate_agregated_experiments_metrics(experiments_starting_time, experiment
 
     metrics_carla.get_all_experiments_aggregated_metrics(result, experiments_starting_time_str, experiments_metrics_and_titles)
     metrics_carla.get_per_model_aggregated_metrics(result, experiments_starting_time_str, experiments_metrics_and_titles)
+    metrics_carla.get_all_experiments_aggregated_metrics_boxplot(result, experiments_starting_time_str, experiments_metrics_and_titles)
 
     with open(experiments_starting_time_str + '/' + 'experiment_elapsed_times.json', 'w') as f:
         json.dump(experiments_elapsed_times, f)
@@ -303,7 +304,8 @@ def main():
             experiments_elapsed_times['total_experiments_elapsed_time'] = time.time() - experiments_starting_time
             generate_agregated_experiments_metrics(experiments_starting_time, experiments_elapsed_times)
     if app_configuration.experiment_random_spawn_point == True:
-        os.remove('tmp_circuit.launch')
+        if os.path.isfile('tmp_circuit.launch'):
+            os.remove('tmp_circuit.launch')
     logger.info('DONE! Bye, bye :)')
